ENSI-BIGDATACRAWLER/app/requirements.txt
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
azure-mgmt-containerinstance==2.0.0
azure-mgmt-resource==8.0.1
azure-storage-file-datalake==12.1.2
Flask==2.0.3
pandas==1.2.1
xlrd==1.2.0
pyarrow==0.16.0
redis==3.4.1
selenium==3.141.0
unrar==0.4
rarfile==3.1
requests==2.23.0
beautifulsoup4==4.9.0
openpyxl==3.0.3
libarchive==0.4.7
cython==0.29.21
python-dotenv==0.15.0
zipfile2==0.0.12
grequests==0.6.0
py7zr==0.16.2
urllib3==1.25.11
zipfile-deflate64==0.2.0
unicodecsv==0.14.1
simpledbf==0.2.6
chardet==3.0.4